Turn off the filtering and display of review status
After show and tell, we think we don't want to use the finder's
capability in its current form (relies on a two-valued facet rather
than there just being a present/absent). But we do need to be
recording this metadata. So, we keep sending it to publishing-api and
Rummager and leave the back-end UI in for it, but remove any
public-facing representation.